some temp hackerfoo to check component_map
The check sees if a component has already been set,
this way we can avoid the hr Prompt and hr Negative
Prompt from overwriting the Prompt and Negative Prompt

Additional work will need to be done so we can include those
as well.

Currently the goal is to get the core plugin to function.